[PATCH] filesystem hashes: NUMA interleaving
The following patch modifies the dentry cache and inode cache to enable the
use of vmalloc to alleviate boottime memory allocation imbalances on NUMA
systems, utilizing flags to the alloc_large_system_hash routine in order to
centralize the enabling of this behavior.

In general, for each hash, we check at the early allocation point whether
hash distribution is enabled, and if so we defer allocation.  At the late
allocation point we perform the allocation if it was not earlier deferred.
These late allocation points are the same points utilized prior to the
addition of alloc_large_system_hash to the kernel.

Signed-off-by: Brent Casavant <bcasavan@sgi.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41da18d5XzbMNt2_2fQkvcSTlf08fw