[PATCH] IDE_ARCH_OBSOLETE_INIT fix
The attached patch fixes the IDE driver to initialise correctly in the case
that IDE_ARCH_OBSOLETE_INIT is not defined.  Not defining this macro would
seem to be the correct thing to do since it includes the word "obsolete" in
its name.

Signed-Off-By: David Howells <dhowells@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41da1e42HRpPg9Pw6Np7FuwJ6jeTUQ