[PATCH] x86: remove data-header and code overlap in boot/setup.S
The setup.S data and code is not cleanly separated.  The current space
reserved for setup header is not enough to fit the complete header.  As a
result, part of the header (EDDBUF) will overwrite the initial code.

I haven't seen any negative impact of this bug.  As, by the time the setup
code is overwritten, we would have finished executing it anyway.  But, I
think it is better to separate the header and code and prevent this
data_overwriting_the_code condition.

The atatched patch adds some space in the header to fit all the data listed
in Documentation/i386/zero-page.txt for both i386 and x86_64 (and updates
zero-page.txt).

Signed-off-by:: "Venkatesh Pallipadi" <venkatesh.pallipadi@intel.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41da1e5cr46H3GyaONfaYkALUeHWKA