[PATCH] x86_64: Add SRAT NUMA discovery to x86-64.
Add SRAT NUMA discovery to x86-64.

Previously the NUMA nodes on Opteron systems were directly read from the
hardware registers of the northbridge.

This version also supports reading it from the more generic ACPI SRAT table.
This allows to support NUMA on all kinds of systems and in general gives more
flexibility, but needs a bit of BIOS support.

This implementation was not done before because the SRAT specification had a
bad license from Microsoft that didn't allow to implement it on other OS.  Now
that ACPI 3.0 is released and SRAT is part of it that isn't a concern anymore.

One side effect of this is that holes not described in SRAT (e.g.  usually the
640K DOS and 4GB PCI memory holes) will not included in the nodes and won't
get mem_map entries allocated.  This may have side effects for some drivers.

Some support for handling dual core AMD CPUs is included.

SLIT parsing is done in a later patch.  Depends on earlier patches to split up
the ACPI boot parsing and to add better CMP support on x86-64.

Can be disabled with numa=noacpi

Signed-off-by: Andi Kleen <ak@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41da1f72IvQWZVQ3P-3fGjJKEgPfcA