[PATCH] uml: redo the signal delivery mechanism
This patch rips out the old signal delivery mechanism and replaces it with a
sane one.

Specifically, UML used to capture a host signal frame and use it as a template
for its own signal frames.  This was a worthy idea, because it promised a
degree of architecture-independence for this part of UML, but impractical.
There are some environments, notably 32 bit emulation on a 64 bit box, where
you can't use the host signal frame as a template for your own.  Plus, this
code is as complicated, even to someone who understands what it's doing, as
the standard fill-in-a-structure-and-write-it-to-the-stack.  For everyone
else, it is incomprehensible.

So, this reimplements signal handling in the way that everyone else does.  It
gives up on architecture independence, and moves this code into the
x86-specific stuff.

Signed-off-by: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41da2470DGQR-pwOnOo8P8FJYTEf5g