[PATCH] uml: kill host processes properly
This patch changes how UML kills ptraced processes in order to be more
correct in the presence of the ptrace changes in 2.6.9.  It used to be that
ptrace stopped processes could simply be killed and they would go away.  Now,
there's a new run state for ptraced processes which doesn't receive signals
until they are PTRACE_KILLed or PTRACE_CONTinued.  So, this patch kills the
process, as usual, then PTRACE_KILL and PTRACE_CONT.  This is done in
os_kill_ptrace_process() for use from skas mode, and in tracer() when it
sees a child process getting a SIGKILL.

Signed-off-by: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41da2556i7p7dyfQRouV4ppFnkxq1g