[PATCH] ppc64: use c99 initializers
This patch is just more clean up in the ppc64 arch.  It uses c99 initializers
for various iSeries structures that are used to pass information to the
hypervisor.  Also itLpNaca is not used by any code that could be in a module,
so don't export it.

Built and booted.

Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41db77ab7l1vcAfSlT_zqp_6G7krHA