[PATCH] frv: Fix PCI config space write
The attached patch makes byte and word writes to PCI config space work.
The problem was that the pointer to the appropriate chunk of the config
port needs to be juggled to allow for the fact that FRV is big endian in
this case.

Signed-Off-By: David Howells <dhowells@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41db78d31MfZO8cXURWrCOsW1TXJqw