[PATCH] Fix usage of setup_arg_pages() in IA64, MIPS, S390 and Sparc64
The attached patch fixes the usage of setup_arg_pages() in the IA64, MIPS,
S390 and Sparc64 arches.  This function now takes an extra parameter: the
initial top of stack.  This is useful in uClinux when there's no fixed
location to which the stack pointer can be initialised.

Signed-Off-By: David Howells <dhowells@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41db794eDCX3fIqwII_Q_5AFuHbaKw