[PATCH] Cleanups for the IPMI driver
This patch removes some unneeded cruft that Adrian found, and also turns
off the shutdown of the timer when removing the module.  Since the timer is
shutdown when the driver is closed (unless no way out is specified) this is
unnecessary and defeats the no way out option.

- remove some completely unused code
- make some needlessly global code static
- removal of some EXPORT_SYMBOL'ed code with zero users.
- Removal of the timer shutdown on module removal

Signed-off-by: Adrian Bunk <bunk@stusta.de>
Signed-off-by: Corey Minyard <minyard@acm.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41db7aa21iXIv9X-gxpxd802qUp16A