[PATCH] prio_tree: move general code from mm/ to lib/
Last but not least, move the general prio_tree code from mm/ to lib/.  This
patch also duplicates some macros, which are used in the VMA code for
debugging purposes, so we can't properly separate them yet.

Note that this patch conflicts with a patch in 2.6.10-rc3-mm1
(frv-better-mmap-support-in-uclinux.patch), which removes mm/prio_tree in
systems without an MMU.  Not making that other patch provide a dummy for
prio_tree_init should resolve the conflict.  (That's just from reading the
patch - I haven't actually tried this.)

Signed-off-by: Werner Almesberger <werner@almesberger.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41db7cf62kQMHxBx7d2wEtYLWrJgKg