[PATCH] SELinux: audit task comm if exe cannot be determined
This patch ensures that the comm is included in the audit message if avc_audit
is unable to determine the exe due to the mmap_sem being held.  This is
helpful in tracking down the causes of permission denials that occur in the
mmap/mprotect hooks.

Signed-off-by:  Stephen Smalley <sds@epoch.ncsc.mil>
Signed-off-by:  James Morris <jmorris@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41db7dc9CpuzE6vOV3F83ORxwcq8og