[PATCH] pm2fb: fix fbi image display on 24 bit depth big endian
Handle 24bit on big-endian by leaving the hardware in RGB and using the
colour offset to reverse red/blue.  Leaving the hardware in RGB means that
fbi displays images correctly (abeit fortuitously) and the console is
correct, thanks to a recent patch to the console code correcting an endian
bug in fbcon_putc().

Signed-off-by: Jim Hague <jim.hague@acm.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41db810f6sKIvyNJ8cI2CYwVFgsOtw