Merge pobox.com:/garz/repo/netdev-2.6/8390 into pobox.com:/garz/repo/net-drivers-2.6
2004/12/27 15:18:56-05:00 penguin
[PATCH] 2.6.9 Use skb_padto() in drivers/net/8390.c

The 8390 driver had been fixed for leaking information in short packets
prior to skb_padto() existing.  This change gets rid of the scratch area on
the stack and makes it use skb_padto().  Thanks to Mark Smith for bringing
this to my attention.

Signed-off-by: Paul Gortmaker <p_gortmaker@yahoo.com>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

BKrev: 41ddbd38ptN2_b1Gc9pKohggbt_1nA