[PATCH] m32r: build fix
This patch is required to fix compile errors for m32r.

This was originally given by the following patch:
  [PATCH] move irq_enter and irq_exit to common code
  http://www.ussg.iu.edu/hypermail/linux/kernel/0411.1/1738.html

I think it was maybe accidentally dropped only for the m32r arch due to a
patching conflict with the other patches or something like that.

Signed-off-by: Hirokazu Takata <takata@linux-m32r.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e052a7WRDa_jwhS9d_W3ReqKE0zQ